add product & sku

+++ boot-showcase-management/src/main/java/org/lina/boot/web/controller/DatatablesViewPage.java
@@ -7,8 +7,8 @@ import java.util.List;
  */
 public class DatatablesViewPage<T> {
     private int draw;
-    private int recordsTotal;
-    private int recordsFiltered;
+    private long recordsTotal;
+    private long recordsFiltered;
     private List<T> data;
 
     public List<T> getData() {
@@ -27,19 +27,19 @@ public class DatatablesViewPage<T> {
         this.draw = draw;
     }
 
-    public int getRecordsTotal() {
+    public long getRecordsTotal() {
         return recordsTotal;
     }
 
-    public void setRecordsTotal(int recordsTotal) {
+    public void setRecordsTotal(long recordsTotal) {
         this.recordsTotal = recordsTotal;
     }
 
-    public int getRecordsFiltered() {
+    public long getRecordsFiltered() {
         return recordsFiltered;
     }
 
-    public void setRecordsFiltered(int recordsFiltered) {
+    public void setRecordsFiltered(long recordsFiltered) {
         this.recordsFiltered = recordsFiltered;
     }
